requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
et-xmlfile==1.0.1
future==0.18.2
idna==2.8
# Editable install with no version control (invest-scripts==0.0.0)
-e /Users/tianli/Desktop/coding_projects/invest_scripts_repo/invest_scripts
jdcal==1.4.1
numpy==1.18.0
openpyxl==3.0.2
pandas==0.25.3
pycparser==2.19
python-dateutil==2.8.1
python-telegram-bot==12.4.2
pytz==2019.3
requests==2.22.0
selenium==3.141.0
six==1.13.0
soupsieve==1.9.5
tornado==6.0.3
urllib3==1.25.7
xlrd==1.2.0
